ROhdsiWebApi-package {ROhdsiWebApi}	R Documentation

ROhdsiWebApi: Interacting With an OHDSI WebApi Instance

Description

ROhdsiWebApi is an R based interface to OHDSI WebAPI services, and performs GET/PULL/POST/DELETE calls via the WebApi. All objects starting from R or output to R - are analysis ready R-objects like list and data.frame. The package handles the intermediary steps by converting R-objects to JSON and vice versa.

Author(s)

Maintainer: Gowtham Rao rao@ohdsi.org

Authors:

Martijn Schuemie

Jamie Gilbert JGilber2@its.jnj.com

Other contributors:

Observational Health Data Science and Informatics [copyright holder]

See Also

Useful links:

https://ohdsi.github.io/ROhdsiWebApi/

https://github.com/OHDSI/ROhdsiWebApi

Report bugs at https://github.com/OHDSI/ROhdsiWebApi/issues

[Package ROhdsiWebApi version 1.3.1 Index]