checkInputFileEncoding {ROhdsiWebApi}	R Documentation

Check character encoding of input file

Description

For its input files, CohortDiagnostics only accepts UTF-8 or ASCII character encoding. This function can be used to check whether a file meets these criteria.

Usage

checkInputFileEncoding(fileName)

Arguments

fileName
The path to the file to check

Value

Throws an error if the input file does not have the correct encoding.

[Package ROhdsiWebApi version 1.3.1 Index]