convertConceptSetDefinitionToTable {ROhdsiWebApi}	R Documentation

Convert a concept set definition to a table

Description

Convert a concept set definition to a table

Usage

convertConceptSetDefinitionToTable(conceptSetDefinition)

Arguments

conceptSetDefinition
A concept set definition, for example as obtained through the getConceptSetDefinition function, or taken from a cohort definition.

Value

Takes a R (list) representation of the Concept Set expression and returns a table (dataframe) representing the concept set expression. This is useful to create publication friendly output of the concept set expression.

Examples

## Not run: 
conceptSetDefinition <- getConceptSetDefinition(conceptSetId = 282,
                                                baseUrl = "http://server.org:80/WebAPI")
convertConceptSetDefinitionToTable(conceptSetDefinition = conceptSetDefinition)

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]