createConceptSetWorkbook {ROhdsiWebApi}	R Documentation

Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook

Description

Save a set of concept sets expressions, included concepts, and mapped concepts into a workbook

Usage

createConceptSetWorkbook(
  conceptSetIds,
  fileName,
  baseUrl,
  included = FALSE,
  mapped = FALSE
)

Arguments

conceptSetIds
A vector of concept set IDs.

fileName
The name of the XLSX workbook file.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

included
Should included concepts be included in the workbook?

mapped
Should mapped concepts be included in the workbook?

Value

A xlsx workbook that includes a list of all concept set IDs and names and a worksheet for the concepts in each set. Options to include an included concepts and mapped concepts worksheet for each concept set are available.

[Package ROhdsiWebApi version 1.3.1 Index]