getCohortDefinitionExpression {ROhdsiWebApi}	R Documentation

(Deprecated) Get a cohort definition expression

Description

(Deprecated) Get a cohort definition expression

Usage

getCohortDefinitionExpression(cohortId, baseUrl)

Arguments

cohortId
An integer id representing the id that uniquely identifies a cohort definition in a WebApi instance.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

Details

(Deprecated) Obtain the JSON expression from WebAPI for a given cohort id

Value

A JSON list object representing the cohort definition This function has been deprecated. As an alternative please use the following steps as in the example below: 1) cohortDefinition <- getCohortDefinition(baseUrl = baseUrl, cohortId = 15873) 2) validJsonExpression <- RJSONIO::toJSON(cohortDefinition$expression) 3) save validJsonExpression object as .txt"

Examples

## Not run: 
# This will obtain a cohort definition's JSON expression:

getCohortDefinitionExpression(cohortId = 282, baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]