getCohortDefinitionName {ROhdsiWebApi}	R Documentation

(Deprecated) Get a cohort definition's name from WebAPI

Description

(Deprecated) Get a cohort definition's name from WebAPI

Usage

getCohortDefinitionName(baseUrl, cohortId, formatName = FALSE)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

cohortId
An integer id representing the id that uniquely identifies a cohort definition in a WebApi instance.

formatName
Should the name be formatted to remove prefixes and underscores?

Details

(Deprecated) Obtains the name of a cohort. This function has been deprecated. As an alternative please use getCohortDefinition

Value

The name of the cohort.

[Package ROhdsiWebApi version 1.3.1 Index]