getDefinitionsMetadata {ROhdsiWebApi}	R Documentation

Retrieve the meta data for WebApi definitions of a certain category

Description

Retrieve the meta data for WebApi definitions of a certain category

Usage

getDefinitionsMetadata(baseUrl, category)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

category
These are the categories in WebApi. The valid string options are 'conceptSet', 'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.

Details

Obtains the meta data of WebApi specifications such as id, name, created/modified details, hash object, etc for a certain category. The following function categories are supported. Concept-set, Cohort-definition, Cohort-characterization, Pathway-analysis, Incidence rate (ir), estimation and prediction. This function is useful to retrieve the current specifications.

Value

A tibble of specification metadata.

Examples

## Not run: 
getDefinitionsMetadata(baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]