getGenerationInformation {ROhdsiWebApi}	R Documentation

Get generation information.

Description

Get generation information.

Usage

getGenerationInformation(id, category, baseUrl)

Arguments

id
An integer id representing the id that uniquely identifies a definition for the category in a WebApi instance.

category
The category of expression in WebApi. Only the following strings are accepted: 'cohort', 'characterization', 'pathway', 'incidenceRate'.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

Details

Get generation (execution) information.

Value

An R object with the generation information.

Examples

## Not run: 
getGenerationInformation(id = 13242,
                         category = "cohort",
                         baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]