getPersonProfile {ROhdsiWebApi}	R Documentation

Get person profile data

Description

Get person profile data

Usage

getPersonProfile(baseUrl, sourceKey, personId, indexCohortId = NULL)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

sourceKey
The source key for a CDM instance in WebAPI, as defined in the Configuration page

personId
The personId of the person whose profile is being reviewed

indexCohortId
<OPTIONAL> Do you want to use a particular cohortId as the index cohort? If left NULL, the WebApi will identify the earliest cohort for the person by cohort start date and use it as the index cohort. WebApi uses the cohort start date of the index cohort to calculate the person's index age (ageAtIndex). WebApi will also return the relative position, in days, for each event compared to the index cohorts start date. These relative positions are useful to study the relationship of various events with respect to the index cohort start date.

Details

Get a R object with person profile data. This function may be used for visualizing a patients profile in tables or visualization.

Value

A list of tibble data frame objects corresponding to cohorts, observationPeriod, records and person.

Examples

## Not run: 
getPersonProfile(baseUrl = "http://server.org:80/WebAPI",
                 sourceKey = "sourceKey",
                 personId = 342342)

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]