invokePathwayGeneration {ROhdsiWebApi}	R Documentation

Invoke generation of Pathway id.

Description

Invoke generation of Pathway id.

Usage

invokePathwayGeneration(pathwayId, baseUrl, sourceKey)

Arguments

pathwayId
An integer id representing the id that uniquely identifies a Pathway definition in a WebApi instance.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

sourceKey
The source key for a CDM instance in WebAPI, as defined in the Configuration page

Details

Invoke the generation of Pathway id in the WebApi.

Value

A tibble with job status information.

Examples

## Not run: 
invokePathwayGeneration(pathwayId = 13242,
                        baseUrl = "http://server.org:80/WebAPI",
                        sourceKey = "HCUP")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]