isValidCohortId {ROhdsiWebApi}	R Documentation

is Cohort id a valid definition in the WebApi.

Description

is Cohort id a valid definition in the WebApi.

Usage

isValidCohortId(cohortIds, baseUrl)

Arguments

cohortIds
A list of integer id(s) of the Cohort to be tested for validity.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

Details

Checks if a set of id for a Cohort is valid. The following checks are performed. 1) checks if all the ids exists in the WebApi i.e. valid.

Value

A logical vector indicating if an ID is valid.

Examples

## Not run: 
isValidCohortId(cohortIds = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]