isValidId {ROhdsiWebApi}	R Documentation

Check if an id is valid.

Description

Check if an id is valid.

Usage

isValidId(ids, baseUrl, category)

Arguments

ids
A list of integer id(s) of the category to be tested for validity.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

category
These are the categories in WebApi. The valid string options are 'conceptSet', 'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.

Details

Checks if a set of id for a category is valid, i.e. checks if all the ids exists in the WebApi i.e. valid.

Value

A logical vector indicating if an ID is valid.

Examples

## Not run: 
isValidId(ids = c(13242, 3423, 34), baseUrl = "http://server.org:80/WebAPI", category = "cohort")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]