postCharacterizationDefinition {ROhdsiWebApi}	R Documentation

Post Characterization definition.

Description

Post Characterization definition.

Usage

postCharacterizationDefinition(name, characterizationDefinition, baseUrl)

Arguments

name
A valid name for the definition. WebApi will use this name (if valid) as the name of the definition. WebApi checks for validity, such as uniqueness, absence of unacceptable character etc. An error might be thrown.

characterizationDefinition
An R list object containing the expression for the specification. This will be converted to JSON expression by function and posted into the WebApi. Note: only limited checks are performed in R to check the validity of this expression.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

Details

Post Characterization definition to WebAPI

Value

This function will return a dataframe object with one row describing the posted WebApi expression and its details. If unsuccessful a STOP message will be shown.

Examples

## Not run: 
postCharacterizationDefinition(name = "new valid name",
                               characterizationDefinition = definition,
                               baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]