Post ConceptSet definition.

Post ConceptSet definition.

postConceptSetDefinition(name, conceptSetDefinition, baseUrl)

Arguments

name
A valid name for the definition. WebApi will use this name (if valid) as the name of the definition. WebApi checks for validity, such as uniqueness, absence of unacceptable character etc. An error might be thrown.

conceptSetDefinition
An R list object containing the expression for the specification. This will be converted to JSON expression by function and posted into the WebApi. Note: only limited checks are performed in R to check the validity of this expression.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

Value

This function will return a dataframe object with one row describing the posted WebApi expression and its details. If unsuccessful a STOP message will be shown.

Details

Post ConceptSet definition to WebAPI

Examples

if (FALSE) {
postConceptSetDefinition(name = "new valid name",
                         conceptSetDefinition = definition,
                         baseUrl = "http://server.org:80/WebAPI")
}