resolveConceptSet {ROhdsiWebApi}	R Documentation

Resolve a concept set to the included standard concept IDs

Description

Resolve a concept set to the included standard concept IDs

Usage

resolveConceptSet(conceptSetDefinition, baseUrl, vocabularySourceKey = NULL)

Arguments

conceptSetDefinition
A concept set definition, for example as obtained through the getConceptSetDefinition function, or taken from a cohort definition.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

vocabularySourceKey
The source key of the Vocabulary. By default, the priority Vocabulary is used.

Details

Resolve a concept set to the included standard concept IDs

Value

A vector of standard concept ids.

Examples

## Not run: 
conceptSetDefinition <- getConceptSetDefinition(conceptSetId = 282,
                                                baseUrl = "http://server.org:80/WebAPI")
conceptIds <- resolveConceptSet(conceptSetDefinition = conceptSetDefinition,
                                baseUrl = "http://server.org:80/WebAPI")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]