setAuthHeader {ROhdsiWebApi}	R Documentation

Manually set the authorization http header for a WebAPI baseUrl In some cases the user may want to manually set the authorization header. An authHeader is associated with a particular baseUrl and added to to the header of all http requests sent to that url by ROhdsiWebApi.

Description

Manually set the authorization http header for a WebAPI baseUrl In some cases the user may want to manually set the authorization header. An authHeader is associated with a particular baseUrl and added to to the header of all http requests sent to that url by ROhdsiWebApi.

Usage

setAuthHeader(baseUrl, authHeader)

Arguments

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

authHeader
A character string containing a Bearer token that will be added to the header of all http requests sent to baseUrl. (e.g. "Bearer lxd9n2nsdsd2329km23mexjop02m23m23mmmsioxiis0")

[Package ROhdsiWebApi version 1.3.1 Index]