updateConceptSetDefinition {ROhdsiWebApi}	R Documentation

Update a ConceptSet definition.

Description

Update a ConceptSet definition.

Usage

updateConceptSetDefinition(conceptSetDefinition, baseUrl)

Arguments

conceptSetDefinition
An R list object containing the expression for the specification. Must include id, name and expression. This will be converted to JSON expression by function and posted into the WebApi. The definition will be checked against the WebApi instance for errors

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

Details

Update a ConceptSet definition.

Examples

## Not run: 
definition <- getConceptSetDefinition(id = 13242,
                                      baseUrl = "http://server.org:80/WebAPI",
                                      category = conceptSet)
definition$name <- "My new name for this"
updateConceptSet(conceptSetDefinition, baseUrl, category = "cohort")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]