updateDefinition {ROhdsiWebApi}	R Documentation

Update definition

Description

Update definition

Usage

updateDefinition(definition, baseUrl, category)

Arguments

definition
An R list object containing the expression for the specification. This will be converted to JSON expression by function and posted into the WebApi.

baseUrl
The base URL for the WebApi instance, for example: "http://server.org:80/WebAPI".

category
These are the categories in WebApi. The valid string options are 'conceptSet', 'cohort', 'characterization', 'pathway, 'incidenceRate', 'estimation', 'prediction'.

Details

Update a definition in WebAPI. Currently only cohorts are supported. Takes the definition as a parameter and converts it to json. This is the full definition (i.e. including name and id fields)

Examples

## Not run: 
definition <- getDefinition(id = 13242, baseUrl = "http://server.org:80/WebAPI", category = "cohort")
definition$name <- "My new name for this"
updateDefinition(definition, baseUrl, category = "cohort")

## End(Not run)

[Package ROhdsiWebApi version 1.3.1 Index]